third_party/blink/web_tests/external/wpt/css/selectors/invalidation/has-sibling-expected.txt

This is a testharness.js-based test.
Found 71 tests; 36 PASS, 35 FAIL, 0 TIMEOUT, 0 NOTRUN.
PASS initial_color
FAIL add .test to first_sibling assert_equals: expected "rgb(0, 128, 0)" but got "rgb(128, 128, 128)"
PASS remove .test from first_sibling
FAIL add .test to second_sibling assert_equals: expected "rgb(255, 0, 0)" but got "rgb(128, 128, 128)"
PASS remove .test from second_sibling
FAIL add .test to third_sibling assert_equals: expected "rgb(255, 0, 0)" but got "rgb(128, 128, 128)"
PASS remove .test from third_sibling
FAIL add .test to first_sibling_child assert_equals: expected "rgb(255, 192, 203)" but got "rgb(128, 128, 128)"
PASS remove .test from first_sibling_child
FAIL add .test to first_sibling_descendant assert_equals: expected "rgb(255, 255, 0)" but got "rgb(128, 128, 128)"
PASS remove .test from first_sibling_descendant
FAIL add .test to third_sibling_child assert_equals: expected "rgb(128, 0, 128)" but got "rgb(128, 128, 128)"
PASS remove .test from third_sibling_child
FAIL add .test to third_sibling_descendant assert_equals: expected "rgb(0, 0, 255)" but got "rgb(128, 128, 128)"
PASS remove .test from third_sibling_descendant
FAIL insert element div.test before first_sibling assert_equals: expected "rgb(0, 128, 0)" but got "rgb(128, 128, 128)"
PASS remove element div.test before first_sibling
FAIL insert element div.test before second_sibling assert_equals: expected "rgb(255, 0, 0)" but got "rgb(128, 128, 128)"
PASS remove element div.test before second_sibling
FAIL insert element div.test before third_sibling assert_equals: expected "rgb(255, 0, 0)" but got "rgb(128, 128, 128)"
PASS remove element div.test before third_sibling
FAIL insert element div.test before first_sibling_child assert_equals: expected "rgb(255, 192, 203)" but got "rgb(128, 128, 128)"
PASS remove element div.test before first_sibling_child
FAIL insert element div.test before first_sibling_descendant assert_equals: expected "rgb(255, 255, 0)" but got "rgb(128, 128, 128)"
PASS remove element div.test before first_sibling_descendant
FAIL insert element div.test before third_sibling_child assert_equals: expected "rgb(128, 0, 128)" but got "rgb(128, 128, 128)"
PASS remove element div.test before third_sibling_child
FAIL insert element div.test before third_sibling_descendant assert_equals: expected "rgb(0, 0, 255)" but got "rgb(128, 128, 128)"
PASS remove element div.test before third_sibling_descendant
FAIL insert element div.test after first_sibling assert_equals: expected "rgb(255, 0, 0)" but got "rgb(128, 128, 128)"
PASS remove element div.test after first_sibling
FAIL insert element div.test after second_sibling assert_equals: expected "rgb(255, 0, 0)" but got "rgb(128, 128, 128)"
PASS remove element div.test after second_sibling
FAIL insert element div.test after third_sibling assert_equals: expected "rgb(255, 0, 0)" but got "rgb(128, 128, 128)"
PASS remove element div.test after third_sibling
FAIL insert element div.test after first_sibling_child assert_equals: expected "rgb(255, 192, 203)" but got "rgb(128, 128, 128)"
PASS remove element div.test after first_sibling_child
FAIL insert element div.test after first_sibling_descendant assert_equals: expected "rgb(255, 255, 0)" but got "rgb(128, 128, 128)"
PASS remove element div.test after first_sibling_descendant
FAIL insert element div.test after third_sibling_child assert_equals: expected "rgb(128, 0, 128)" but got "rgb(128, 128, 128)"
PASS remove element div.test after third_sibling_child
FAIL insert element div.test after third_sibling_descendant assert_equals: expected "rgb(0, 0, 255)" but got "rgb(128, 128, 128)"
PASS remove element div.test after third_sibling_descendant
FAIL insert tree div>div.test before first_sibling assert_equals: expected "rgb(255, 192, 203)" but got "rgb(128, 128, 128)"
PASS remove tree div>div.test before first_sibling
FAIL insert tree div>div.test before second_sibling assert_equals: expected "rgb(128, 0, 128)" but got "rgb(128, 128, 128)"
PASS remove tree div>div.test before second_sibling
FAIL insert tree div>div.test before third_sibling assert_equals: expected "rgb(128, 0, 128)" but got "rgb(128, 128, 128)"
PASS remove tree div>div.test before third_sibling
FAIL insert tree div>div.test before first_sibling_child assert_equals: expected "rgb(255, 255, 0)" but got "rgb(128, 128, 128)"
PASS remove tree div>div.test before first_sibling_child
FAIL insert tree div>div.test before first_sibling_descendant assert_equals: expected "rgb(255, 255, 0)" but got "rgb(128, 128, 128)"
PASS remove tree div>div.test before first_sibling_descendant
FAIL insert tree div>div.test before third_sibling_child assert_equals: expected "rgb(0, 0, 255)" but got "rgb(128, 128, 128)"
PASS remove tree div>div.test before third_sibling_child
FAIL insert tree div>div.test before third_sibling_descendant assert_equals: expected "rgb(0, 0, 255)" but got "rgb(128, 128, 128)"
PASS remove tree div>div.test before third_sibling_descendant
FAIL insert tree div>div.test after first_sibling assert_equals: expected "rgb(128, 0, 128)" but got "rgb(128, 128, 128)"
PASS remove tree div>div.test after first_sibling
FAIL insert tree div>div.test after second_sibling assert_equals: expected "rgb(128, 0, 128)" but got "rgb(128, 128, 128)"
PASS remove tree div>div.test after second_sibling
FAIL insert tree div>div.test after third_sibling assert_equals: expected "rgb(128, 0, 128)" but got "rgb(128, 128, 128)"
PASS remove tree div>div.test after third_sibling
FAIL insert tree div>div.test after first_sibling_child assert_equals: expected "rgb(255, 255, 0)" but got "rgb(128, 128, 128)"
PASS remove tree div>div.test after first_sibling_child
FAIL insert tree div>div.test after first_sibling_descendant assert_equals: expected "rgb(255, 255, 0)" but got "rgb(128, 128, 128)"
PASS remove tree div>div.test after first_sibling_descendant
FAIL insert tree div>div.test after third_sibling_child assert_equals: expected "rgb(0, 0, 255)" but got "rgb(128, 128, 128)"
PASS remove tree div>div.test after third_sibling_child
FAIL insert tree div>div.test after third_sibling_descendant assert_equals: expected "rgb(0, 0, 255)" but got "rgb(128, 128, 128)"
PASS remove tree div>div.test after third_sibling_descendant
Harness: the test ran to completion.
